Move all clipboard components into single folder
