+++ packages/common/src/components/Clipboard/TextboxClipboardWrapper/styles.tsx
